+++ src/utils/Constant.js
@@ -1,7 +1,6 @@
 const MESSAGE = Object.freeze({
   GAME_START: '다리 건너기 게임을 시작합니다.\n',
-  GAME_RETRY:
-    '\n게임을 다시 시도할지 여부를 입력해주세요. (재시도: R, 종료: Q)\n',
+  GAME_RETRY: '\n게임을 다시 시도할지 여부를 입력해주세요. (재시도: R, 종료: Q)\n',
   READ_BRIDGE_SIZE: '다리의 길이를 입력해주세요.\n',
   READ_MOVE_LEVEL: '\n이동할 칸을 선택해주세요. (위: U, 아래: D)\n',
 });
@@ -41,10 +40,7 @@ const OUTPUT_MESSAGE = Object.freeze({
 });
 
 const format = (message, ...args) => {
-  return args.reduce(
-    (formatted, arg, index) => formatted.replace(`{${index}}`, arg),
-    message,
-  );
+  return args.reduce((formatted, arg, index) => formatted.replace(`{${index}}`, arg), message);
 };
 
 module.exports = {
